[build-utils] Remove duplicated node backend code (#14031)
We have lots of duplicated code to handle `express`, `hono`, `h3`, and
`nestjs`. This PR merges duplicated code in `build-utils`.

Only `h3` and `nestjs` are updated with this PR since `express` and
`hono` are currently being worked on for o11y. A future PR will also
refactor `express` and `hono` to use this once o11y work is more stable.

---------

Co-authored-by: vercel[bot] <35613825+vercel[bot]@users.noreply.github.com>