Add vc dev support for NestJS (#14080)
NestJS entrypoints may look like:
```ts
async function bootstrap() {
  const app = await NestFactory.create(AppModule);
  await app.listen(process.env.PORT ?? 4321);
}
bootstrap();
```

which means we can't assume `.listen` was called immediately like
`compileUserCode` used to expect. In `compileUserCode` now we handle the
server case last, and if `.listen` was not called by the point, we wait
up to a second for the server listen to happen.