[@vercel/firewall] Add optional env var for path prefix to use with Firewall rate limit requests (#14123)
In order to support microfrontends or applications using rewrites, this
PR adds the ability for users to set an optional path prefix to the
Firewall Rate Limit request endpoint. The path prefix would be used to
route to the correct microfrontend, and the child microfrontend would
then need to remove the path prefix via a rewrite. The
`@vercel/microfrontends` package will automatically apply this logic so
users don't need to think about it.