[build-utils]: Remove getSpawnOptions (#14176)
# Overview

Remove the `getSpawnOptions` function from being used for subprocesses.
This function is designed to setup any subprocesses running from the CLI
with the correct version of node in `PATH` for the build. We've ported
this logic to the build environment now so that any processes running
are run with the right `PATH` environment variable set for their build
settings (respecting `package.json`).

I've refactored the builders to now only use `getEnvForPackageManager`
in places which require the subprocess. The intention here is to remove
"vercel environment specific" logic from the CLI and have that owned by
the build container - so that the CLI is a smaller wrapper around
builder specific logic.

This logic is next to be moved to the Vercel environment so that the
right zero-config detection is run when spawning the CLI, which makes
this logic detection redundant.