Use ts 5 for backends only (#14252)
Some frameworks like Elysia and Fastify fail during `vc build` if
there's no Typescript version installed because the internal types are
relying on `>5.x`.

We need to upgrade TS across the board
([pr](https://github.com/vercel/vercel/pull/14205)), but for now, just
do it for the backend builders so we don't affect `api` functions.

---

Also notable, this PR removes typechecking during `vc dev`. Most modern
build tools aren't doing it here and if/when we move to rolldown we
won't be doing it then, either.