[next]: only produce data routes when using pages router (#14235)
We currently generate routing rules for `/_next/data` routes even if
there is no pages router directory present. If you had a bunch of
dynamic routes that had `generateStaticParams`, we'd create routes for
each of these even though they would never be utilized. Vercel has
limitations to the number of routes your application can have, so this
is wasteful and can easily push an app over the current maximum.

Depends on:
- https://github.com/vercel/next.js/pull/86051

Fixes NAR-477