Fix token expiry check (#14306)
We currently check if an OIDC token is within 15 minutes of expiration
which creates a race condition when using the new OIDC client in
production. If a token expires at T100, we may check at T90 if it is
expired and return true. This causes a refresh which then fails as there
is no local token.

Fix logic to not check with a delta. Clients of this should not be
caching a token, even for the duration of a function, so checking exact
expiry vs now is correct.