[CICD-DX] Improve project selection with input selector (#14280)
Improve project linking experience in CLI.

If a user has <= 100 projects for their team, we allow users to scroll and search for their projects. For users with more than 101, we maintain the current flow where users need to type the project in directly. Supporting users with a large number of project names requires more network calls, and may not necessarily be a notably improved experience. This threshold can be reevaluated in the future.

A v2 of this feature could include a search or scroll-through-all-projects, though would introduce more network calls. For users with few projects, a input selector view seems preferable though.

### <100 projects

![CleanShot 2025-11-14 at 15.18.43@2x.png](https://app.graphite.com/user-attachments/assets/9a89356a-3e0c-4755-bdc0-d6998c194788.png)

### \> 100 projects

![CleanShot 2025-11-14 at 15.20.15@2x.png](https://app.graphite.com/user-attachments/assets/8940272c-a7fd-4536-a4f0-aee524bee77a.png)

### No projects

![CleanShot 2025-11-14 at 15.21.12@2x.png](https://app.graphite.com/user-attachments/assets/605f96ed-b0ba-4a7e-a726-46697522197e.png)  
  
ref https://linear.app/vercel/issue/FLOW-4715/make-project-names-autocomplete-when-linking-to-existing-project