[python] pyproject build commands fix (#14303)
Running a build command from `pyproject.toml` previously wrapped
commands with `uv run` but this introduced a couple issues mentioned
below. This PR simplifies the `runPyprojectScript` function logic so
commands are executed directly without being wrapped with `uv run`.

**Issues**
When running build commands with `uv run ...`, uv automatically looks
for a `.python-version` file. If it is found, it creates a .venv virtual
environment and, if the requested python version is not available,
installs it on the fly.

This means we can run build commands with arbitrary python versions but,
since a .venv for the requested python version was also created, `uv pip
install ...` will install dependencies for the requested python version
instead of the deployment python version which will crash the function
since python won't be able to find the necessary dependencies.