[CICD-DX] prefer first production alias (#14307)
We've encountered a use-case where when targeting rolling releases the
prior behavior that picked a more specific deployment URL means that
rolling releases would therefore be disabled.

This PR fixes that by defaulting to a URL (if one is not provided with
`--deployment`) that supports that use-case. In a way, this is also,
perhaps, a more reasonable default in general.