[cli] create util to execute and set config path conditionally based on vercel.ts (#14312)
Created a new util to conditionally check whether a `vercel.ts`
configuration file exists, and if so return it as the config path.
Existing logic to check between a `now.json` and a `vercel.json` is
still available.

This util executes the TypeScript configuration file and generates an
intermediary `vercel.json` inside `VERCEL_DIR`. It's protected by a
feature flag, and will be protected similarly in the actual commands
where it's used (`dev`, `build`, and `deploy`).