[CICD-DX] Include the branch name in output for `vc pull` (#14351)
Include the branch name in output for `vc pull`. 

Improves messaging when using `vc pull` with a specified preview environment, as well as a branch to indicate that variables will be pulled for the environment AND the branch provided.

`vc pull --environment=preview --git-branch=this-is-a-branch`

fixes https://linear.app/vercel/issue/FLOW-1003/improve-message-on-vercel-pull-environment=preview-when-git-branch