[client] Add retry logic to deployment status polling (#14407)
This is an attempt to fix
https://vercel.slack.com/archives/C09MK639NMN/p1764806810028989?thread_ts=1761256638.364529&cid=C09MK639NMN

We do a lot of deployments in parallel in the Next.js CI, and hit the
rate limit very often:

https://github.com/vercel/next.js/actions/workflows/test_e2e_deploy_release.yml

This is a simple GET request (safe to retry), and the API returns an
HTTP 429 in this case, so we should just use the `Retry-After` header. I
also added a check for HTTP 5xx errors.

I used claude to help write some unit tests, though I cleaned up and
simplified what claude generated.

The rate limit for this endpoint is pretty high (500/min), but I tested
this manually by removing the normal `getPollingDelay` sleep at the end
of the generator loop, and by adding a bunch of artificial parallel
fetches to the code.