[cli] Add commands for bulk redirects feature (#14393)
This PR adds new commands for `vercel redirects` to interact with the
upcoming bulk redirects feature on Vercel. This feature allows users to
manage redirects at the project level, without needing to change
redirects from a deployment.


https://github.com/user-attachments/assets/2a826f71-765c-400f-8f26-0d91c61b6ced

The PR adds the following commands:

- `list` - lists redirects with support for versions, searches, and
pagination
- `list-versions` - lists all redirects versions
- `promote` and `restore` - for updating the live version
- `add` and `remove` for adding and removing redirects

The CLI commands will return `Error: Bulk redirects feature is not
enabled for this team.` until this feature is enabled.

---------

Co-authored-by: Austin Merrick <onsclom@onsclom.net>