Skip introspection on routes that fail path-to-regex conversion (#14423)
We use `path-to-regex` to convert Hono routes to a regular expression.
They don't map 1:1 and in the case of a `/*` route, a failure on the
conversion caused all o11y to fail.

The proper fix here will be to map hono routes to their appropriate
regex logic, which isn't an API they've exposed, as far as I can tell.