[next]: fix clobbering of dynamic route data (#14468)
`00-app-dir-no-ppr` in newer versions of Next.js was catching a logic
bug in the builder that omitted important details on the `route` when
reconstructing it. I removed the failing assertion in
https://github.com/vercel/vercel/pull/14465 just so we can get tests
running again.

In this case, when client segment cache is enabled, RSC requests to
`fallback: false` pages like `/nested/blog-fallback-false/non-existent`
in the test app match the dummy RSC output (`{}`) rather than the
correct RSC route instead of falling through to the app router
catch-all.

I unpinned `00-app-dir-no-ppr` since there's no reason for this test app
to be pinned to an older version of Next.js.