[cli] Add vc redirects upload command for bulk upload of redirects (#14459)
This PR adds a `vc redirects upload` command to upload CSV or JSON files
to Vercel.

<img width="1063" height="400" alt="Screenshot 2025-12-11 at 10 33
05 AM"
src="https://github.com/user-attachments/assets/95407814-be01-4784-99af-6f040bec5d08"
/>

The CLI will do some light sanity checks on the files in the client and
then upload them to the server. If successful, it will then fetch the
diff for the new redirects in that upload and display them to the user.
Similar to the `add` command, it can then prompt the user to promote if
it is the only set of staged redirects.