Update codeowners for config package (#14508)
Code owners for the config package have a few problems:

- Missing cicd (all others have it)
- Includes edge-content and edge-security

This PR updates the code owners to remove the extra teams, re-add cicd,
and add myself specifically.

As a follow-up, we will need someone to add the `@vercel/edge-ingress`
team as a collaborator on `vercel/vercel` so that that team can be
recognized by GitHub.